requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pillow==11.2.1
pillow-heif==1.3.0
pydantic-settings==2.9.1
python-multipart==0.0.20
mangum==0.19.0
